Parser `CalcJobNode` fixture: add inputs before storing node
The `generate_calc_job_node` fixture for the parser tests mocks a
calculation job node, but it was failing, because it was adding inputs
after having stored the `CalcJobNode`. A recent change in `aiida-core`
forbids inputs to be added after a `ProcessNode` is stored. The fix is
to simply add the input nodes, before storing the process node itself.
Since at least one of the nodes has to be stored when adding a link, the
input nodes now have to be explicitly stored, if they weren't already.