Create a single CLI script `aiida-quantumespresso`
This new executable, installed through the `console_scripts` entry point
supersedes all the individual launcher scripts that were being added.
Having a single script for each calculation or workflow launching script
was not tenable as it was polluting the namespace of the environment.

The new `aiida-quantumespresso` has various advantages:

 * Only a single executable is required to host all CLI commands
 * Name makes immmediately clear that it belongs to this plugin
 * The `click` library is used to make a robust CLI tool
 * Tab-completion can be enabled with single line in `.bashrc`

Note that the `fastentrypoints.py` script is used to make the loading of
the entry points a lot faster. Without it, loading the tool takes 2
seconds instead of 0.2 seconds. Without it, tab-completion is unusable.