Fix bug `NamelistCalculation._get_following_text` and clean-up (#293)
The function signature of `_get_following_text` specified that the
settings dictionary should be passed, but it wasn't. Instead of passing
it, we remove it from the signature, because the only class that
overrides the method `MatdynCalculation` does not use it. If it will be
necessary in the future, it can be re-added.