Implement pytest fixtures to efficiently test `CalcJob` classes
The same principle is used as for the parser test fixtures. Various
fixtures are provided to easily create sample inputs, such as structures
kpoints and upf nodes. Then the main fixture `generate_calc_job` should
be called with a temp folder and the inputs, which will then instantiate
the process for those inputs and call `prepare_for_submission`. The test
itself then receives the `CalcInfo` returned by that method, which it
can then inspect, as well as the contents that have been written to the
sandbox folder as the raw inputs of the calculation. This allows to test
`CalcJob` implementations as unit tests without actually having to run
them through the engine.