Add a test for the `PwParser` for a `relax` calculation
Note that since the raw output parser depends explicitly on the input
parameters, for example to determine what type of calculation was run,
the inputs need to contain the correct calculation type or the parser
will not parse all the output nodes. That is why the `relax` test
requires a separate input node generator function, in order to be able
to set the correct `calculation` type in the input parameters.