`PwParser`: refactor the handling of error messages
Currently, both the standard output and the XML output are parsed. The
current approach is based on older versions of Quantum ESPRESSO where
the XML output was a sub set of that of the standard output. Therefore
the XML was parsed first and then passed to the function that parses the
stdout, using the XML output to complement where possible. With the
recent addition of a schema-based XML output, were slowly moving towards
making the XML output leading.

It becomes very important to make the parsing of these two output
sources as independent as possible. The `PwParser` can then decide how
to merge the parsed data from both sources and which should take
precedence. As well it should decide when and which error codes to
return. Ideally, before exiting, it should parse as much information as
possible.

This commit makes a start with refactoring the parser and the raw
parsers to separate these tasks as well as possible.