Remove the outdated `examples` module
These examples were heavily outdated and people keep finding them and
trying to run them with `v3.0.0`. The examples also tried to fullfill
two roles in the way they were written: as test scripts and as example
CLI scripts for users. The former is now done much better through actual
pytest unit tests, without having the need to run the binary. The second
use is now also replaced by a binary `aiida-quantumespresso` installed
with the plugin that provides a single consistent CLI.

The only thing that will have to be replaced is that actual examples on
how to build and run a calculation should be added to the documentation.