`BaseRestartWorkChain`: improve and adapt to `aiida-core==1.0.0`
The base restart work chain was originally developed to be able to deal
with problems in the submission or retrieval of calculation jobs to the
remote clusters. However, with the addition of the exponential back off
retry mechanism for calculation jobs in the engine of `aiida-core` this
no longer needs to be dealt with by the restart work chain. The logic of
the calculation is inspected. In addition, the position of the sanity
check invocation is moved from failed calculations to finished
calculations. The whole concept of the sanity check is to add an
additional check on calculations that were marked as succesful by the
parser. Calculations that were already marked failed do not have to have
a sanity check.