`PwBaseWorkChain`: document spec and fix automatic parallelization (#326)
The spec did not specify any help messages yet for the inputs and the
outputs which have now been added. Additionally, the automatic
parallelization functionality has been fixed. There were a few calls to
methods that have been removed in `aiida-core==1.0.0`. Note that this
functionality requires an initialization run, which is performed by
running `pw.x` with an exit file in the directory. This will cause the
code to perform the preamble and then exit. This functionality seems to
be broken for versions of `pw.x` that produce the XML output with the
new schema format. As a result the automatic parallelization only works
for codes that produce the XML in the old format.