`BaseRestartWorkChain`: check for existence of `_error_handlers`
The `WorkChain._error_handlers` attribute is only initialized when a
handler gets registered through the `register_error_handler` decorator.
This is because it cannot be done on the `BaseRestartWorkChain` base
class since then handlers will be attached to all sub classes and not
just the one to which the decorator is applied. However, this means that
if *no* handlers are registered, the class will have no `_error_handlers`
attribute. The `_handle_calculation_failure` method, however, was
assuming its existance, resulting in an `AttributeError`. The fix is to
simply explicitly check for the existence of the attribute first.