Add triggered error handlers as extra to the process node
The `register_error_handler` decorator is updated to automatically
attach the name of a handler to the work chain node's `errors_handled`
extra, if it returns an `ErrorHandlerReport`, meaning an error was
handled. This allows one to query for work chains that had certain
errors handled during their execution. The `errors_handled` extra will
be a list of lists, where each entry in the list represents a
calculation run by the work chain, since more than one can be called,
and the values of each list are the names of the error handlers that
were triggered for that particular calculation.

To benefit from this and the other features of the `register_error_handler`
decorator, one also might want to define a method to handle certain errors,
but only call it manually and not have it automatically included in the
`handle_calculation_failure` loop. To accomplish this, we make the
`priority` argument of the decorator optional, which will exclude it
from being called automatically, however when called, will still trigger
the addition of its name to the `errors_handled` as well as the
automated reports in case of a `verbose` run.