`BaseRestartWorkChain`: abort due exceeded iterations in right place
It is currently done in the `inspect_calculation` after the check if the
calculation is finished ok. However, if the sanity_check returns an
`ErrorHandlerReport` the loop is restarted, but if the sanity check failed
on the calculation of already the last iteration, the outline would
break out of the while loop and go to the results. Here it will attempt
to attach the results from the `restart_calculation` to the outputs of the
work chain, but that has been set to `None` by the `sanity_check`
resulting in an exception being thrown.

The solution is to move the check to the `results` step. If the number
of iterations has been exceeded, the work chain will always end up here.
We then only have to check if the last calculation has failed and the
number of iterations has been exceeded, in which case the corresponding
exit code is returned, without attaching the outputs, since they most
likely are missing or incorrect.