`PwCalculation`: only ever create `output_trajectory`
The parser would return either the `output_array` or the
`output_trajectory` node, depending on the calculation type. However,
this was making analysis difficult, because one always had to check
which node was attached. On top of that, some of the arrays in the
`TrajectoryData` were moved to the `output_parameters` if it only
contained one value or "frame". This meant yet another check in analysis
code to try and find the data that could be in two places. All these
complications are removed by only ever creating the `output_trajectory`
node which will always contain all the same arrays. For improved
queryability, some of these arrays will have their last value copied
over to the `output_parameters` node as well.