Updated the pw documentation to the latest AiiDA 1.0 version
I checked that the two documentation pages for pw are up to date with
the latest version of AiiDA, including an example script. Modifications inspired from
the latest AiiDA tutorial pages.

Fixed the automodule documentation pages to remove references to outdated classes where it was OK to do so.

Still 2 warnings in the compilation for me, that I did not touch,
reference to non-existing AiiDA classes -- aiida.tools.dbexporters.tcod_plugins