`PwBandStructureWorkChain`: fix various minor bugs
In addition to the fixes for `PwBandStructureWorkChain`, its spec and
that of the `PwBandsWorkChain` and `PwRelaxWorkChain` are updated with
help messages for their input and output ports.

Additionaly, with the recent addition of the feature `populate_defaults`
argument of the `PortNamespace` the `relax` option of the
`PwBandsWorkChain` is now properly optional. Before, the validation
would fail if the `relax` namespace was not defined at all, which is the
mechanism to let the user indicate that the relaxation of the input
structure should be skipped.