Add `qe-tools` dependency
This allows to simplify the parsers in `aiida_quantumespresso.tools` that
have been largely moved to the `qe-tools` package, which is an `aiida-core`
independent package. The parsers cannot be completely removed, because
some `aiida-quantumespresso` specific functionality needed to be added
that was removed from `qe-tools` because it depends on `aiida-core`.
Specifically, the parsers here extend those of `qe-tools` with the methods
`get_structuredata` and `get_kpointsdata` which parse `StructureData`
and `KpointsData` nodes, respectively, from the input file. This
functionality was removed in `qe-tools==1.1.3` to finally make it fully
`aiida-core` agnostic.