Restructure the `aiida_quantumespresso.cli`
The sub commands need to be defined in their proper sub modules and
there is no need for a separate `cli` module within the package. The
definition of the root command can be placed directly in the module.

Also moved an import in `cli.utils.launch` within the function
definition to prevent the ORM being loaded and slowing down the loading
time during tab completion.