Fix and simplify the calculation restart builder utility (#378)
A lot of the functionality that was implemented in the restart utilities
in `aiida_quantumespresso.utils.restart` has become obsolete by the
`CalcJobNode.get_builder_restart` method which already returns a
`ProcessBuilder` instance with its inputs populated based on the source
node. The utility then merely needs to adapt some inputs such as setting
the correct restart parameter and `parent_folder` node.