`MatdynCalculation`: change `parent_folder` to `force_constants` (#387)
The `MatdynCalculation` inherited the `parent_folder` input port from
the `NamelistCalculation` base class, except the name is a bit
unfortunate. All other namelist calculations indeed take a parent folder
as input, which can either be a remote or retrieved folder data.
However, for the `MatdynCalculation` this should actually be a
`ForceConstantsData` single file node, produced by a `Q2rCalculation`.