Import physical constants from `qe-tools` instead of `aiida-core` (#388)
The physical constants, mostly used by the parsers, used to be defined
in `aiida-core` for the Quantum ESPRESSO plugins, before they were moved
to this repository. The physical constants remained, however, but will
be removed for `aiida-core==1.0.0`. The constants had already be copied
to `qe-tools` as well, which is where they will stay. Here we change the
import to use the constants from the latter instead of `aiida-core`.