Soften empty band occupation requirements in `get_highest_occupied_band`
The utility function, in addition to determining the HOMO, also checked
that all bands above it are "empty" as well, since, when not the case, it
is a hint of non-sensical band structures. However, the check was that
once the LUMO was found, all subsequent bands had to have an occupation
that was lower. This gave problems for band structures that have
negative occupations just below the fermi level. This can happen with
both Methfessel-Paxton and Marzari-Vanderbilt smearing. The LUMO would
be marked as the band with negative occupations and any subsequent band
with occupation of zero, would trigger the exception. Although
physically meaningless, negative occupations are well-known effects of
these smearing methods and need to be accepted. The solution is to
change the criterion to compare unoccupied bands not to the occupation
of the LUMO but to the specified threshold. As long as all unoccupied
bands have an occupation less than twice the threshold (where the factor
two accounts for numerical noise) the criterion of bands above the HOMO
being empty is preserved within the bounds of numerical noise.