`PwParser`: fix sorting bug in parser of new XML format (#411)
The `fft_grid` and `fft_smooth` fields in the `output_parameters` were
built from the dictionary returned by the raw parser. The dictionary
was turned into a list without taking the order into account, however
the ordering of the dictionary do have meaning. To guarantee a correct
and consistent order we sort the dictionary before building the list.