`NebParser`: simplify tests and add check on `ArrayData` attributes
The input generator for the parser tests only need to specify those
nodes that also will be accessed by the parser class. Additionally, an
explicit test is added on the attributes of the `ArrayData` and
`TrajectoryData` output nodes which contain information on what arrays
have been parsed and their shapes.