`NebParser`: remove `output_trajectory` content from regression
The content of the `TrajectoryData` arrays `cells` and `positions` is
parsed by the base parser of `PwParser` and represent the cells and
positions of the atoms. Since this parser is already tested elsewhere
and the data required for the regression file is large, we remove it
from an explicit check on the content. The attributes of the trajectory
node are verified so that guarantees that all arrays are there with the
correct shape.