Add exit code 551 for 'negative dexx' errors
In calculations with hybrids, a common error is that the code
aborts with a 'negative dexx' message. Possible causes can be an
incorret 'exxdiv_treatment', or insufficient 'ecutfock', but the
error is also sometimes mysterious. Here, a specific exit code
is added to the PwCalculation class for this error, to enable
handling in the workflows.