`NebParser`: reduce symmetries in the output parameters
The `output_parameters` node contains the symmetries as determined by
Quantum ESPRESSO in each image. These dictionaries are quite big and
repetitive and unnecessarily bloat the database. The `PwParser` already
implemented functionality to reduce the information on a simple mapping.
That functionality is abstracted to a stand alone function such that the
`NebParser` can also call it.

The content of the fixture file `neb/default/aiida.int` is also reduced
significantly. This is simple interpolated data that is loaded as a
whole by the parser into a numpy array. Checking 10 values is just as
good as checking a 1000.