Added SCF convergence trajectory to the pw parser
Adressing suggestions by @sphuber

The pw parser now look for the SCF convergence in each SCF cycle of
a relaxation. The data is saved in two arrays of the output_trajectory.

scf_accuracy is a 1D array containing all the estimated errors
found in the output;
scf_accuracy_index acts as an index to identify different relax steps
in scf_accuracy. Each entry is the index of the FIRST item of a specific relax step;
e.g. the relaxation convergence in the second step might be extracted to
the array xscf as:
  i1 = scf_accuracy_index[1]
  i2 = scf_accuracy_index[2]
  xscf = scf_accuracy[i1: i2]

The separation is two array is to avoid inhomogeneous arrays, see
aiidateam/aiida-core#3411

Parsing of these quantities will allow for more complex logic in the workflows