`PwParser`: allow parsing of incomplete XML output with new format
The parser of the XML output file through the new schema-based parsing
assumed the output would always be complete. However, for example during
an initialization run, where only the preamble of the code is run, this
is not the case and certain information specified in the schema as
required is not actually parsed. Since this is a valid use of `pw.x` our
parser should be able to deal with this.

The schema based XML parser will throw a lot of errors because of the
incomplete tags as it does not check whether it was an initialization
run. To prevent these verbose messages from floading the log, all parser
error messages are suppressed.