SCF convergence trajectory to pw parser, setdefault instead of try; except
The pw parser now look for the SCF convergence in each SCF cycle of
a relaxation. The data is saved in two arrays of the output_trajectory.

scf_accuracy is a 1D array containing all the estimated errors
found in the output.

Replaced the try; catch instances of dictionary construction by the
setdefault method within the section