Migrate `NebCalculation` and `NebParser` to `aiida-core==1.0.0` (#381)
The `NebCalculation` shares a lot of functionality of `PwCalculation`,
since in essence running a `neb.x` calculation involves running various
`pw.x` calculations under the hood, but not quite enough to make sub
classing it from `BasePwCpInputGenerator` the correct choice. Instead
we use `expose_inputs` that expose the relevant inputs ports of the
`PwCalculation` class and used the `_generate_PWCPinputdata` to create
the `pw.x` related input. For this purpose said method was transformed
into a class method.

A similar approach is used for the `NebParser` which leverages most of
the `PwParser` functionality to parse the stdout put of each image. Also
here some methods of `PwParser` were transformed into static or class
methods. The functionality to reduce symmetries through mapping, that
was an instance method of `PwParser`, has been moved to free standing
functions in the `aiida.parsers.parse_raw.pw` module so it could be
used by the `NebParser` as well