Enable checking of log messages in parser unit tests
The parsers configure a logger on the calculation node, which will
automatically insert `Log` entries in the database when a message is
emitted through that logger. Tests for successful calculations now check
that there are no log messages, whereas tests for failed calculations
verify the opposite, that there are in fact logs attached to the node.

Note that currently the `Log.objects.get_logs_for` method does not
provide a way to filter for log levels. Once this is implemented in
`aiida-core` the tests should be updated to only check for `warning` and
or `error` messages.