`PwParser`: add the SCF accuracies to the `TrajectoryData` (#407)
The `PwParser` will now parse the SCF accuracy values from all SCF
cycles and store them in the `TrajectoryData` output node in two arrays:

  * `scf_accuracy`
  * `scf_accuracy_index`

The accuracy values of all SCF cycles are stored as a one-dimensional
array in `scf_accuracy`. This is because the length of each SCF cycle
can be different and `ArrayData` nodes do not allow inhomogeneous arrays
because the pose a security risk, due to them requiring pickles to store
them to disk. To keep the knowledge of what parts of `scf_accuracy`
belong to what SCF iteration the `scf_accuracy_index` keeps an array
with the starting index of each SCF cycle within the `scf_accuracy`
array. The last element corresponds to the index of the last entry in
the `scf_accuracy` array for easy processing.

Additionally, a case is added to the raw pw output parser that will
catch cases where the calculation is stopped during an SCF iteration.