`PwCalculationTools`: fix bug in `get_scf_accuracy` (#426)
The implementation was incorrect when receiving negative indices. In
addition the implementation now uses the `scf_iterations` array instead
of the `scf_accuracy_index`. The former has existed for a long time and
the latter was recently added, but since the information is redundant
and takes up unnecessary space, it is removed from the `TrajectoryData`.