`PwParser`: check for specific exit codes before generic one (#439)
The `PwParser` would throw the generic `ERROR_OUTPUT_FILES` if both the
parsers of the stdout as well as of the XML returned an error. However,
this would prevent more specific error codes, often available from the
stdout parser, to be obscured. We switch the order of the error checking
and first consider specific errors that are known to be reported in the
stdout. Only if that fails, do we fall back on the more generic error
code `ERROR_OUTPUT_FILES` if both parsers reported an error.