Enable CI on Github through actions
Add a Github workflow that will run on `push` and `pull_request`. It
will build the docs and run `pre-commit` on python 3.7 only. The tests
are run through `pytest` on a matrix of python 2.7, 3.6 and 3.7. Caching
of the pip installation is enabled and seems to work, as in it gets hit,
however it does not seem to result in a reduced installation time. The
origin of this is unknown as of yet.