`PwParser`: add new exit code `ERROR_OUT_OF_WALLTIME_INTERRUPTED`
If the `max_seconds` parameter is set in the input and this number is
exceeded, Quantum ESPRESSO will attempt to terminate the calculation
cleanly by writing all files to disk that are necessary for a restart.
However, during this phase the scheduler can still kill the job if the
total requested walltime has been exceeded. This will result in the
restart files being corrupt and not suitable for a restart. Therefore
this case needs a different exit code than the currently existing one
`ERROR_OUT_OF_WALLTIME` which should only be used for calculation that
shutdown cleanly.