`PwParser`: add support for QE v6.5 (#460)
Version 6.5 of Quantum ESPRESSO comes with a new XML schema version for
the pw.x output. We add the corresponding schema file `qes_191206.xsd`
which then automatically allows the `PwParser` to parse the output of
this newer version. The existing test `default_xml_new` which tested the
older `qes_190304.xsd` schema has been renamed to `default_xml_190304`
such that a new one could be added `default_xml_191206` for the new
schema.