`generate_calc_job_node`: use structure from input file if present
Certain parsers will expect the output with respect to the crystal
structure to be commensurate with the input structure. So far the
`generate_calc_job_node` was using a default silicon structure as input.
This was fine for the initial error cases for which the reference output
files were easy to generate. As these become more complicated one does
not want to be forced to reproduce it with silicon if one already has an
example reference file. Instead it would be useful to have the fixture
just get the input structure by parsing it from the input file. The
`generate_calc_job_node` is updated to do just this: it will try to
parse the structure from the `aiida.in` file in the fixture folder and
if successful override the `structure` in the inputs.