`PwParser`: check for ionic convergence before BFGS history fail
Quantum ESPRESSO can stop with the BFGS history reset error even when
all forces and stresses are already below the thresholds. This happens
when the structure is already relaxed. For some reason QE will still
enter the optimization loop and often BFGS will struggle. In this case,
the parser returned `520` but this should really just be considered
as relaxed and so return `0`. The solution is to, in the case of the
presence of the BFGS warning message, to first manually check the
convergence of the forces and optionally the stress.

Note that the reference files for `vcrelax_failed_bfgs_history` had to
be adapted manually. This already existant test that was supposed to
test the `ERROR_IONIC_CYCLE_BFGS_HISTORY_FAILURE` exit code, was also
accidentally victim of the bug described above and so after these
changes would now be marked successful. To use it for its original
purpose the pressure was increased manually such that the convergence
test would fail.