Make trajectory validation utils raise instead of return `None`
The utility functions in `aiida_quantumespresso.utils.validation.trajectory`
would return `None` if one of the requested arrays or indices within it
do not exist. This made the return value difficult to use, because the
`None` could easily be confused with a `False`. If a non-existing array
or index is specified it is better to raise an exception, so now the
functions raise a `ValueError`.