`PwParser`: use input parameters for ionic convergence thresholds
The `validate_ionic` method was accidentally using the output parameters
which does not contain the convergence thresholds parameters and so the
default values were always used, ignoring the actual values used in the
inputs of the calculation.

To bring this bug to light, we first replace the reference files for the
`vcrelax_success` test with those of a run with convergence thresholds
that are much higher than the default. Due to the bug, the parser would
mark the calculation incorrectly as unconverged, but the change in this
commit fixes that.