`PwBandsWorkChain`: make `nbands_factor` optional
By making this a non required input, the caller has to opportunity to
explicitly specify a number of bands to be used for the bands step
through the `bands.pw.parameters` input node. One can now specify the
number of bands to be used in the BANDS step either directly through
the input parameters `bands.pw.parameters.SYSTEM.nbnd` or through
the `nbands_factor` input. Note that specifying both is not allowed.
When neither is specified nothing will be set by the work chain and
the default of Quantum ESPRESSO will end up being used. If the
`nbands_factor` is specified the maximum value of the following values
will be used:

  * `nbnd` of the preceding SCF calculation
  * 0.5 * nspin * nelectrons * nbands_factor
  * 0.5 * nspin * nelectrons + 4 * nspin