`seekpath_structure_analysis`: expand parameter input into keywords
The original design required a single `Dict` node that contained all the
parameters that were to be passed to Seekpath. However, since arbitrary
keyword arguments are now supported in process functions it is better to
expect individual data nodes. Since all parameters for the Seekpath
function will expect simple data types, such as a string, int, float or
bool, we unpack the nodes in the kwargs into their underlying base
values and pass it on.