`PwBandsWorkChain`: make kpoints for bands configurable
The caller may want to define an explicit kpoint data node to use for
the bands calculation. This can now be passed as `bands_kpoints` in
which case the Seekpath part is completely skipped because if the
structure is changed in that operation the input kpoints may no longer
be congruent. Instead, if `bands_kpoints` is not defined, Seekpath is
called and the returned explicit kpoint path will be used. The input
`bands_kpoint_distance` can be used to change the default reference
distance defined by Seekpath to control the minium distance between
k-points in reciprocal space.