Homogenize the exit codes for all `CalcJob` implementations
Making sure that all calculation jobs use the same exit codes for the
same problem does not only make things easier for the user but it can
then also be used in code to make certain assertions based on the exit
code of a failed calculations. An example is the automatically
invalidating of caching for exit codes smaller than 400, which will be
implemented soon. This is only possible if all calculations respect the
convention that codes below 400 are irrecoverable errors and so those
should also not be used for caching.