Automatically invalidate cache of `CalcJob` exit codes below 400 (#453)
The current convention for the definition of exit codes on `CalcJob`
implementations is that irrecoverable errors should get an exit status
smaller than 400. These errors are also those errors that should remove
a calculation from being considered in the caching mechanism. The new
`invalidates_cache` attribute of the `ExitCode` tuple, introduce in
`aiida-core==1.1.0` allows us to mark exit codes that should disqualify
a calculation from being considered.

Thanks to the convention on exit statuses that we maintain for all calc
job implementations, we can automate this rule by providing a specific
version of `CalcJob` that has a `ProcessSpec` class that will
automatically set `invalidates_cache` to `True` if the status is below
400. The only criterion for this to work is that all `CalcJob` implemen-
tations now use this base class instead of the one provided by aiida-core

Co-Authored-By: Sebastiaan Huber <mail@sphuber.net>