`PwParser`: allow parsing from only the stdout
This is useful in cases where the XML output file is not available, for
example in older calculations run without AiiDA that need to be
imported. We define a new option on the `PwCalculation` class called
`without_xml` that is not required. We do not set a default otherwise
that would be set as an attribute on all nodes. If it is set to `True`
in the inputs, a missing XML output file will not cause the parsing to
fail as in the default case.

For certain calculation types, such as `relax` one of the required
outputs is the `structure`. Since the data for this output node is
nominally parsed from the XML data, this would normally fail. However,
we can use the input structure in this case, simply replacing the atomic
positions and optionally lattice vectors that are also parsed from the
stdout. Note that we do not do this in the raw stdout parsing function
because doing so would couple it more tightly to input data being passed
in through arguments, which is something we want to prevent.