`PwBandsWorkChain`: carry updated number of bands between steps
The `PwBaseWorkChain` has a sanity check to make sure the highest bands
are not occupied, unless we are running an insulator without smearing.
If this condition is not met, the number of bands are increased and the
calculation is restarted. We need to make sure that when this happens,
the new number of bands is carried to the next step, because otherwise
the next `PwBaseWorkChain` is most likely going to run in the same
problem. The `PwRelaxWorkChain` was already properly carrying the number
of bands so here we just need to make sure to update the value before
the scf ands bands steps.