`PwParser`: add `smearing_type` and `smearing` from new XML output (#445)
These keys were not yet being parsed from the output file with the new
XML schema. In addition, if an exception is thrown by the raw XML or
stdout parser, it is now caught and added as a critical warning. Finally
the test input generator for `PwParser` is simplified and only a single
generator is used for both scf and relax tests.

Note that the units of the `degauss` depends on the version of the XML
output. It should always be in Hartree, but up until 6.5, it was
erroneously being written in Rydberg. This was corrected  in commit
`a9698a6971e4ad975ce89a15bc9b74294799ef3f`.