`PwBaseWorkChain`: do not restart from scratch after bands sanity check (#492)
The base work chain has a sanity check that will make sure the highest
band has a negligible occupation. If this is not the case, unless the
calculation was explicitly ran with fixed occupations (indicating that
the system is insulating), the number of bands in the inputs is
increased and the calculation is restarted from scratch.

When first implemented, I erroneously assumed that one cannot restart
from the previous calculation because the number of bands will have
changed and so the old files will be incommensurate. This is the case
for certain parameters being changed, such as the wavefunction cutoffs,
but for the number of bands it is perfectly fine to restart. Therefore
we change the behavior of the bands sanity check to *not* restart from
scratch.

This also fixes a bug encountered when this sanity check was triggered
for non-scf calculations. There the restarts would crash horribly
because restarting from scratch in a non-scf calculation makes no sense.
Quantum ESPRESSO would try to read the files from the previous run, that
of course are not there, and would fail miserably.