Add `PpParser` for the `PpCalculation` plugin (#428)
One challenge of pp.x calculations is that there is a choice of both
dimensionality and of output format. As we want to produce AiiDA
`ArrayData` output nodes, the `PpCalculation` plugin is modified to
enforce only Gnuplot (for 1D and 2D) and Cube (3D only) file formats,
based on the dimensionality the user wants. The `PpCalculation` class is
still lightweight in the sense that the user skill is still required to
run pp.x and hand-holding is minimal, but is improved over the previous
'free-form' input version in that the output will definitely be parsed
by AiiDA and stored in the database in a standard way. The parser
collects the useful data from standard out and detects common problems.
or convenience `PpCalculation` also enforces that the post-processed
data is written to a file which is then retrieved and parsed, rather
than to stdout. The parser converts this, for any dimensionality into
the appropriate `ArrayData` representation.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>