Various updates in `setup.json`, `README.md` and CI configuration (#516)
Update the `README.md` to add the prospective release to the
compatibility matrix. It will require `aiida-core>=1.2.0` in order to
add support for `aiida-sssp`.

The `dev` extra requires has been split into `tests` and `pre-commit`
and their dependency requirements have been updated and change to use
the `~=` compatibility operator which is less restrictive. The split
also allows to use more modern version of `pre-commit` for the
corresponding CI step, which dropped Python 3.5 support already, while
still running the unit tests with Python 3.5 while we still support it.

Finally, added some additional pre-commit hooks:

 * double-quote-string-fixer
 * end-of-file-fixer
 * fix-encoding-pragma
 * mixed-line-ending
 * trailing-whitespace

Note that there are no excluded files for these as they would have to be
defined individually for each step. Instead, we just bite the bullet and
accept the many changes to non-critical files now, for simplicity and
consistency in the future.