Replace builtin `BaseRestartWorkChain` with the one from `aiida-core` (#519)
As of `aiida-core==1.1.0` it ships with the `BaseRestartWorkChain` built
in. This version is based directly on the original concept developed in
this plugin, with some minor interface changes and improvements. Here we
remove the old implementation and depend on the version that ships with
`aiida-core`. The base workchains are updated to respected the new
interface and to profit from the new features.

The dependency requirement is updated to `aiida-core~=1.2`, since anyway
soon we will have to rely on features introduced in that version. This
also allows us to simplify the backwards compatibility code that was put
in the base `CalcJob` class for this plugin.

Since this is a big change in the code, not having unit tests for the
work chains was really starting to show. This commit adds some basic
tests for the four base workchains that are currently implemented. They
test that the workchain can succesfully be constructed and `setup` can
be called and almost all of the process handlers are tested. This is
done both by directly calling the handler, as by going through the
`inspect_process` outline step.