Allow explicitly specifying 'ibrav'.
If an 'ibrav' value other than zero is specified in the
SYSTEM inputs, the cell of the input structure is converted
into the appropriate A, B, C, cosAB, cosAC, cosBC values. As a
check, the cell is reconstructed using qe_tools. The cells are
compared element-wise, with an absolute tolerance controlled by
the 'IBRAV_CELL_TOLERANCE' setting.