Remove `prospector` and run `yapf` and `pylint` from repo
The `prospector` tools has been giving a lot of problems with dependency
incompatibilities since we have been using it and we really only use it
to run `pylint`. However, this is easy to run directly as well from a
remote, specifying the desired revision.

We do the same for `yapf` and remove the explicit dependency from the
`pre-commit` extra since the latter was not pinning a particular version
which could cause arbitrary style changes in different environments that
just happen to have a slightly different version.

Note that, although we also install `pylint` directly from a revision
in the pre-commit hook, we still have to define an explicit version in
the `extras_require`. This is because `pylint` needs to be run with the
`system` language defined in the hook, or it won't find all third party
dependencies that are installed and throw a lot of false positives about
missing imports. We define `~=2.5.0` which allows different patch
versions, which should not introduce incompatible linting.